requirements.txt
# LangChain Core
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.39,<0.4.0
langchain_community

# OpenAI Integration
langchain-openai>=0.2.0,<0.3.0
openai>=1.0.0,<2.0.0

# Anthropic Integration
langchain-anthropic>=0.3.0,<0.4.0

# Google Gemini Integration
langchain-google-genai>=2.0.0,<3.0.0
google-generativeai>=0.8.0,<1.0.0

# Hugging Face Integration
langchain-huggingface>=0.1.0,<0.2.0
transformers>=4.40.0,<5.0.0
huggingface-hub>=0.20.0,<1.0.0

# Environment Variable Management
python-dotenv>=1.0.0

# Machine Learning Utilities
numpy>=1.24.0
scikit-learn>=1.3.0

# Groq Integration
langchain-groq



streamlit>=1.20.0

faiss-cpu>=1.7.4

pinecone-client>=1.1.1
pymilvus>=2.2.0
PyPDF2>=3.0.0
pytest>=7.0.0
tqdm>=4.0.0


